Fixed some bugs and Added "Log-In" screen
App now opens on an opening screen with a drop down menu to select the
institution and a field for the hex code.

App passes selection and hex code to Question Activity and this decodes
both, creating the IP address.

Currently does not connect automatically. Networking starts on answer
selection as SocketException occurs.

Needs to be checked with the presentation system now.